unai: de-template inner functions
The inner functions are marked as inline, and are called from templated
functions; this means that the template parameter can be passed as a
regular parameter, as the compiler will still know its value.

De-templating inner functions will later allow to create de-templated
render functions, that will be used as fallback to new optimized
templated functions. The point of de-templating the old ones is then to
avoid exponential growth of the binary size.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>